// src/api/pharmacy/models/OrderStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Waiting List Api
 * Pharmacy Management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xmikova@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Status of the medication order
 * @export
 */
export const OrderStatus = {
    Created: 'created',
    Confirmed: 'confirmed',
    Delivered: 'delivered',
    Cancelled: 'cancelled'
} as const;
export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export function OrderStatusFromJSON(json: any): OrderStatus {
    return OrderStatusFromJSONTyped(json, false);
}

export function OrderStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderStatus {
    return json as OrderStatus;
}

export function OrderStatusToJSON(value?: OrderStatus | null): any {
    return value as any;
}